requirements.txt
praw
openai
tiktoken
langchain
langchain-community
langchain-openai
faiss-cpu
sentence-transformers
tqdm
python-dotenv
